tambahin fungsi login logout(baru untuk mhs sama dosen, untuk kta belum)

+++ src/main/java/com/rpl/project_sista/repository/MahasiswaDashboardRepository.java
@@ -4,6 +4,6 @@ import com.rpl.project_sista.model.entity.TugasAkhir;
 import com.rpl.project_sista.model.entity.Sidang;
 
 public interface MahasiswaDashboardRepository {
-    TugasAkhir findTugasAkhirByMahasiswaId(Long mahasiswaId);
+    TugasAkhir findTugasAkhirByMahasiswaId(int mahasiswaId);
     Sidang findSidangByTugasAkhirId(Long taId);
 }
